pyproject.toml
[project]
name = "ucf-chain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.33.0",
    "opencv-python>=4.12.0.88",
]
